Simplify and centralize sb mapping removal
Whenever we remove a mapping from smc, we also need to remove the
necessary data from data store. Now all the mapping removal are in
seperate places in mapping system making it hard to track whether all
the mapping removal call removing all the data properly.

The goal of the patch is to simplify mapping removal and centralize
all the mapping removal calls in few methods which will remove the
mapping properly from everywhere.

Change-Id: Ia279c019a186e3cdc349eb912abd499bde3e3802
Signed-off-by: Shakib Ahmed <sheikahm@cisco.com>