Bug 7272: Fully remove mappings
When removing a mapping, remove all subkeys, and the radix trie entry,
to avoid false lookup results.

Also, instead of sending a notification w/o a subscriber list, expecting
the listeners to look up the subscribers from the mapping system,
include the subscriber list with the notification. That way all subkeys
can be deleted when the change is a mapping removal.

Change-Id: Ic922b82fab42f533f9fa8ec828e4d21f6fa35198
Signed-off-by: Lorand Jakab <lojakab@cisco.com>