Update patch set 1
Patch Set 1:

(1 comment)

I looked into what happens with Optional<T>#get() when the value is absent: it throws an IllegalStateException, so that's not acceptable. But I found another method, which makes things more readable and robust too: or(T defaultValue). You can specify a default value to use when the Optional is absent. I think this is great, since you don't have to check for either null nor .isAbsent() and can clearly say what happens when the value is absent.

Patch-set: 1