Bug 7947: Store MappingOrigin in MappingData
As a first step towards separate subscriber storage, based on origin,
this patch adds support for storing the origin of a mapping in the
MappingData object. The origin is defined as a Set, to support the use
case of returning a combined mapping with information from different
origins, such as the NB_AND_SB policy surrently.

This patch doesn't change behavior yet, it just stores the origin.

Change-Id: Id285e829fed794de9ecf40ee9a9cb5e775228e92
Signed-off-by: Lorand Jakab <lojakab@cisco.com>