Bug 7947: Store MappingOrigin in MappingData
As a first step towards separate subscriber storage, based on origin,
this patch adds support for storing the origin of a mapping in the
MappingData object. The origin is defined as a Set, to support the use
case of returning a combined mapping with information from different
origins, such as the NB_AND_SB policy surrently.

This patch doesn't change behavior yet, it just stores the origin.

Change-Id: I930fd101dc2b7ca495a9454997d968111901a1b3
Signed-off-by: Lorand Jakab <lojakab@cisco.com>