Move IT back to Karaf 3
Since we're shipping Carbon with Karaf 3, move our integration tests
(IT) for Carbon back to using Karaf 3. Also, don't build the Karaf 4
local distribution, since it takes a lot of time. Additionally, make IT
logs be at TRACE level for lispflowmapping classes, to make it easier to
track down the root cause of potential issues.

Change-Id: I3c0bc0adc394e47c04fac9474e0f9795556f3936
Signed-off-by: Lorand Jakab <lojakab@cisco.com>