Cancel ongoing SMR futures on new event
The SMR scheduler keeps retrying the sending of SMRs a few (configurable
number) of times with a given spacing between the retries until an
SMR-invoked Map-Request is received for the particular EID. When that
happens, the retrying is cancelled for the subscriber that sent it.

However, up until now, when a new set of SMRs was scheduled for a given
EID, the old set, if still retrying, was left untouched. This patch
fixes that, and starts with a clean slate for the given EID.

Change-Id: Ic64ce0980b39d780433ac2c4e96608af12302392
Signed-off-by: Lorand Jakab <lojakab@cisco.com>