Make 'keys' CLI output user friendly
Before this patch, the 'keys' CLI command would print out the full
Java objects contained in a map-cache. That's great for in-depth
debugging, but for a quick look it very hard to read. This patch
implements a user friendly CLI output for keys, to improve
readability.

The old full output is still available in Karaf, by adding the '-d' or
'--debug' switch to the CLI command.

Example output:

---8<------------------------------------------------------------------------------
opendaylight-user@root>keys
Instance ID 0
  -> EID                                           HMAC Algorithm        Shared Key
     192.0.2.1/32                                  HmacSHA1              password
     192.0.2.1/32|192.0.2.2/32                     HmacSHA1              password
------------------------------------------------------------------------------>8---

Change-Id: Ibe95d0ee4318053df3bc9f30c09c0e01c1d00eba
Signed-off-by: Lorand Jakab <lojakab@cisco.com>