Increase some of the integration test timers
After the Jenkins infrastructure was moved to a different cloud,
integration tests started failing in different ways, even after VMs were
upgraded to 4 vCPUs. This patch is a result of trial and error to see
what timings can compensate for the slowdown in MD-SAL performance on
the new infrastructure.

This patch was not needed on my local machine for the build to pass.

Change-Id: I96dc4b8b1f5c3b09df2c24a492e2bcaf0c838114
Signed-off-by: Lorand Jakab <lojakab@cisco.com>