Update patch set 1
Patch Set 1:

I haven't seen an email on this on any of the mailers that I follow, including release. Aren't we supposed to keep following non-SNAPSHOT releases of odlparent? Or is this patch just for some multi-patch test and is never intended to be merged?

Patch-set: 1
Reviewer: Gerrit User 491 <491@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0