Keep integrationtest runtime folder
It is very helpful to have the full Karaf environment used for the
integrationtest stay available after the test finishes. It lives in the
integrationtest/target/exam/{guid} folder, and has the full (~350MB)
Karaf distribution.

The most useful piece is the data/log/karaf.log file, which can be
examined in case of a failure. Other useful pieces are the configuration
files in etc/ which are modified by the integrationtests.

The patch makes sure to delete any existing such runtime environment, to
avoid developers building up many such folders, which takes up a lot of
space, and makes it harder to determine which environment was the last
one.

When running builds in Jenkins, only the log files are saved to the log
server, so the impact is minimal, but it helps debugging integration
test failures which occur on the CI infrastructure.

Change-Id: I11e92efebf6d74c95e4ddf12d02c4ba98462d8ba
Signed-off-by: Lorand Jakab <lojakab@cisco.com>
(cherry picked from commit 2f872ec34363481a93bbcb164f121ce7530ef7a2)