Update patch set 4
Patch Set 4:

> Recently I have added a bundle Util (still in code review), where I
 > think this config fits perfectly in it, as of now. But, there is a
 > scope of config bundle to grow dependency on other bundles which is
 > against the policy of Util, which we plan to keep as a library
 > which can be utilized by other modules.

I agree with this, in fact I had the same idea. But snce this patch is already done, and the util patch is still under discussion, I don't mind merging as-is, and move to util later, to unblock Miro. So good point!

Also, ConfigIni, should not become dependent on other projects, since the idea is to just read the properties files, and provide sane defaults to other bundles.

Patch-set: 4