Update patch set 4
Patch Set 4:

Recently I have added a bundle Util (still in code review), where I think this config fits perfectly in it, as of now. But, there is a scope of config bundle to grow dependency on other bundles which is against the policy of Util, which we plan to keep as a library which can be utilized by other modules.

Patch-set: 4
Label: Verified=0