Update patch set 17
Patch Set 17:

First of all, I don't see any logic for handling the insertion of positive mappings that overlap negative ones. Probably the best way to go about this is to add the positive and remove the negative altogether thereby generating SMRs for all subscribers. This would naturally lead to the indirect pushing of the new positive to the right subscribers and the generation of only the needed negative prefixes for the subscribers falling outside the positive. 

Second, another thing that is missing is the logic to expand negative prefix holes on positive mapping removals. For instance, if a positive is removed, it may be adjacent to a negative mapping/prefix so an aggregate negative should be created. Otherwise, the SMR invoked map-requests for the removal of the positive will just result in a new negative being created whereas our goal is to have as few as possible negative mappings.

Patch-set: 17
Label: Verified=0