Update patch set 1
Patch Set 1:

This is good, but we still need tests with the other address types (mostly the LCAF types), also for the getURIString() and getURLString() calls. For the latter we only need to check with Ipv4Prefix and IPv6Prefix address types, since the rest is not modified for URLs.

Patch-set: 1
Label: Code-Review=+2