Update patch set 13
Patch Set 13:

(10 comments)

I like reviewing patches where the more complicated new part is in one place (not spread everywhere) and the rest of the changes is mostly wiring, so thanks for that! ;)

While we're waiting for the manual tests, I have a few relatively minor comments. I didn't review the integration tests yet though.

Patch-set: 13