Update patch set 7
Patch Set 7:

Alright, so it seems we have a timing difference and not everything is done by the time things are about to be cleaned up. The problem seems to be DataStoreBackEnd, which is not synchronizing on txchain access.

Alternatively MappingServiceIntegrationTest.cleanUP() is not properly syncing threads, or something similar. Those 2 second sleeps seem a bit fishy..

Patch-set: 7