Deprecate mappingservice.neutron
It has been many releases since this code has been tested against
OpenStack, and no one is around to maintain it. Additionally, some
architectural concerns were raised in relation to the use of some YANG
models in this module.

Deprecate in the current release and remove in the next release.

Change-Id: I8fe637ff32efb92de6b0f9f99e4c5dcd153a6fe7
Signed-off-by: Lori Jakab <lojakab@cisco.com>